[~] fixed i2c

+++ mss_example/SoftConsole/mss_top_sb_MSS_CM3/mss_top_sb_MSS_CM3_app/Helpers/sys_helper/sys_helper.c
@@ -2,8 +2,8 @@
 
 void delay
 (
-	uint8_t time,
-	uint8_t frequency
+	uint32_t time,
+	uint32_t frequency
 )
 {
 	int i = 0;
